api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "takojakos",email = "dziki.sledz@op.pl"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-allauth (>=65.10.0,<66.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "dj-rest-auth[with-social] (>=7.0.1,<8.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]

[tool.poetry]
packages = [{include = "api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
